pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "graphcast-africa"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "cdsapi>=0.6",
    "ecmwf-opendata>=0.3",
    "earthkit-data>=0.9",
    "cfgrib",
    "eccodes",
    "xarray",
    "numpy",
    "netCDF4",
    "jax[cpu]",
    "dm-haiku",
    "dm-tree",
    "graphcast @ git+https://github.com/deepmind/graphcast.git",
    "pyyaml",
    "python-dotenv",
    "tqdm",
    "multiurl",
]

[project.optional-dependencies]
gpu = ["jax[cuda12_pip]"]
dev = ["pytest", "pytest-cov", "black", "ruff"]

[project.scripts]
graphcast-africa = "scripts.run_forecast:main"

[tool.ruff]
line-length = 88

[tool.ruff.lint]
# This enables us to use CPY001: copyright header check
preview = true
# This enables us to use the explicit preview rules that we want only
explicit-preview-rules = true
# all rules can be found here: https://docs.astral.sh/ruff/rules/
extend-select = ["E501", "W", "I", "CPY001", "PGH", "RUF", "TID252"]
ignore=[
    # do not assign a lambda expression, use a def
    "E731",
    # do not use variables named 'l', 'O', or 'I'
    "E741",
    # E721 gives many false positives.
    # Use `is` and `is not` for type comparisons, or `isinstance()` for
    # isinstance checks
    "E721",
    # We don't care much about F841.
    # Local variable ... is assigned to but never used
    "F841",
    # some RUF rules trigger too many changes
    "RUF002",
    "RUF003",
    "RUF005",
    "RUF012",
    "RUF015",
    "RUF021",
    # https://docs.astral.sh/ruff/formatter/#conflicting-lint-rules
    "W191",
    "E111",
    "E114",
    "E117",
    "D206",
    "D300",
    "Q000",
    "Q001",
    "Q002",
    "Q003",
    "COM812",
    "COM819",
]

[tool.ruff.lint.flake8-copyright]
notice-rgx = "\\#\\ Authors:\\ The\\ africlimate\\ AI\\ team\\\r?\\\n\\#\\ SPDX\\-License\\-Identifier:\\ MIT"

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"


[[tool.mypy.overrides]]
module = "requests.*"
ignore_missing_imports = true
